The Bloody Gate
This is a project I made in Unity from scratch using blender to make the map and the whole environment and Unity to simulate all of the animations, enemy AI, and everything else. It is a project I made to learn Unity and I uploaded it on itch.io which has gotten over 1000+ downloads and 15,000+ impressions on the game. https://action7.itch.io/the-bloody-gate